moved all redux files to state folder and changed route of imports in homepage and product page
